Add files via upload
In this task, the objective was to utilize SQL to recover information that answers a particular trade address. The assignment included distinguishing the fitting table containing the significant information and developing an SQL inquiry to extricate the desired data. To start, I carefully looked into the dataset given, which included data such as client points of interest, item postings, and more. Following, I decided that the information related to items, counting their costs, would likely be put away in a table named 'Products.' Taking after the informational given, I made an SQL inquiry utilizing the SELECT articulation to recover all areas ('*') from the 'Products' table. I at that point executed this inquiry in my SQL environment, which brought about in showing all the item information from the table.